requirements.txt
dotenv
pyqt5
python-i18n
pytz